Removed KafkaDataStreamer and refactored KafkaStreamer
